refactor(app): 통합 테스트 import 경로 업데이트 및 ADR 추가
WebSocketIntegrationTestSupport / RoomWebSocketTestSupport 이동에 따라
:app 통합 테스트의 import를 갱신하고, 게임 스케줄러 오버라이드 전략을
ADR 0016으로 문서화한다.
