+++ backend/app/src/test/java/coffeeshout/fixture/GameWebSocketIntegrationTestSupport.java
@@ -1,0 +1,8 @@
+package coffeeshout.fixture;
+
+import coffeeshout.global.config.IntegrationTestConfig;
+import org.springframework.context.annotation.Import;
+
+@Import(IntegrationTestConfig.class)
+public abstract class GameWebSocketIntegrationTestSupport extends RoomWebSocketTestSupport {
+}
