+++ backend/app/src/test/java/coffeeshout/websocket/docs/WsCatalogControllerTest.java
@@ -1,9 +1,8 @@
 package coffeeshout.websocket.docs;
 
-import coffeeshout.support.TestContainerSupport;
-import coffeeshout.support.test.IntegrationTest;
-import coffeeshout.websocket.docs.WsCatalog;
-import coffeeshout.websocket.docs.WsCatalogController;
+import static org.assertj.core.api.Assertions.assertThat;
+
+import coffeeshout.support.IntegrationTestSupport;
 import org.assertj.core.api.SoftAssertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
@@ -14,11 +13,8 @@ import org.springframework.context.annotation.Profile;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 
-import static org.assertj.core.api.Assertions.assertThat;
-
-@IntegrationTest
 @DisplayName("WsCatalogController")
-class WsCatalogControllerTest extends TestContainerSupport {
+class WsCatalogControllerTest extends IntegrationTestSupport {
 
     @Autowired
     private TestRestTemplate restTemplate;
