+++ backend/app/src/test/java/coffeeshout/websocket/ui/WsRecoveryControllerTest.java
@@ -4,9 +4,9 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-import coffeeshout.fixture.IntegrationTestSupport;
-import coffeeshout.websocket.WsRecoveryService;
+import coffeeshout.fixture.RoomWebSocketTestSupport;
 import coffeeshout.websocket.StompSessionManager;
+import coffeeshout.websocket.WsRecoveryService;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -16,7 +16,7 @@ import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMock
 import org.springframework.test.web.servlet.MockMvc;
 
 @AutoConfigureMockMvc
-class WsRecoveryControllerTest extends IntegrationTestSupport {
+class WsRecoveryControllerTest extends RoomWebSocketTestSupport {
 
     private static final String TEST_JOIN_CODE = "T3ST";
     private static final String TEST_PLAYER_NAME = "Tester";
