+++ backend/app/src/test/resources/application-test.yml
@@ -1,6 +1,8 @@
 spring:
   config:
     import: classpath:application-test-base.yml
+  main:
+    allow-bean-definition-overriding: true
   datasource:
     url: jdbc:h2:mem:testdb;MODE=MySQL;DATABASE_TO_LOWER=TRUE;DEFAULT_NULL_ORDERING=HIGH
