+++ backend/docs/adr/index.md
@@ -21,3 +21,4 @@
 | [0013](0013-domain-module-test-isolation.md)                  | 도메인 모듈 테스트 독립 실행 전략                                                           | 적용됨                   | `*/src/test/`, `:test-support`, `build.gradle.kts` (testcontainers 버전)                                                                      | TestApplication + TestConfig 패턴; `:app` 고유 빈은 모듈별 TestConfig에서 mock; 공통 설정은 `application-test-base.yml`에 집중 |
 | [0014](0014-shared-web-module.md)                             | 공유 웹 인프라 모듈 `:web` 도입                                                          | 적용됨                   | `RestExceptionHandler`, `SwaggerConfig`, 전 도메인 모듈 `build.gradle.kts`, 루트 `build.gradle.kts`                                               | `:web`은 `:infra` 위에 위치; REST 엔드포인트 모듈은 `:web` 의존; `web`, `validation`, `springdoc`은 `:web`이 `api`로 노출 |
 | [0015](0015-test-fixtures-for-domain-module-fixture-sharing.md) | `java-test-fixtures`를 통한 도메인 모듈 fixture 공유 전략                                 | 적용됨                   | `:room`, `:game`, `:user`, `:app` `build.gradle.kts`, `*/src/testFixtures/`                                                                   | fixture는 의존 도메인 타입을 소유한 모듈의 `testFixtures`에 귀속; cross-cutting fixture는 `:app:test`에 유지; `:test-support`로 이동 불가 (순환 의존) |
+| [0016](0016-game-integration-test-flow-scheduler-override.md) | 게임 통합 테스트 FlowScheduler 실제 구현 주입 전략                                           | 적용됨                   | `app/src/test/`, `test-support/config/BaseIntegrationTestConfig`, `application-test.yml`                                                      | 게임 WebSocket 통합 테스트는 `GameWebSocketIntegrationTestSupport` 상속; `BaseIntegrationTestConfig` mock은 DI 플레이스홀더로 유지 |
